requirements.txt

azure-identity==1.17.1
langchain==0.2.10
langchain-community==0.2.9
openai==1.37.0
requests==2.32.3
tqdm==4.66.4
astunparse

